SFX and Music Plotted
Reviewed the script and marked places for SFX, music, and ambient tracks
with hashtags so that when we have the requisite audio files, we can
locate these spots easier and put them in quickly. This is pending on
final music selections, so they're ambiguous until a later date.

Also misc script fixes that I spotted here and there.